Hardware Acc. & renamed images url for webp
When i press the elevator button on mobile, it gets really slow and choppy when scrolling up. I found out it is because of the image filtering we are doing on the way up. Using the three webkit lines to force hardware acceleration definitely helps a lot.

Note: In the future, i might replace the film images with new inverted/filtered images so that i can reduce the load on the browser. That way, i do not have to rely on hardware acceleration to get my webpage to work. Ive tested the elevator button when the film do not have css filters applied to them, and no hardware acceleration and it runs fine. Its only when i have both the backdrop filter, and image filters for the film section in which things get slow.