fixed previous update. HUGE responsive changes
I was having a weird issue after i included the meta tag in the html head. I noticed that my phone was not consistent with what i was seeing in dev mode. I realized that certain breakpoints were not being met on my phone, and i thought it was an error with the meta tag. but it was no error, its because they really were not being met. I had to overhaul the responsiveness of the webpage so that it looks better on mobile. this is the mobile fitst approach babyyyy. Now, by using the meta tag i am getting the proper device viewport width.